refactor(data): extract shared base for in-memory map data sources
InMemoryJSONDataSource, InMemoryYAMLDataSource and InMemoryTOONDataSource
each repeated the same ~60-line row-iteration block verbatim (fieldsMap and
mapIterator state plus open/nextRow/hasMoreData/reset/iterator/getColumnNames/
readAll). Pull that boilerplate into a new AbstractInMemoryMapDataSource so
each subclass only implements its format-specific parse() into fieldsMap.

No behavior change; the moved method bodies are identical.
