+++ data/src/main/java/io/github/adamw7/tools/data/source/file/InMemoryJSONDataSource.java
@@ -1,33 +1,23 @@
 package io.github.adamw7.tools.data.source.file;
 
 import java.io.InputStream;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import io.github.adamw7.tools.data.source.interfaces.InMemoryDataSource;
-
-public class InMemoryJSONDataSource extends AbstractFileSource implements InMemoryDataSource {
-	private final Map<String, String> fieldsMap = new HashMap<>();
-	private Iterator<String> mapIterator;
+public class InMemoryJSONDataSource extends AbstractInMemoryMapDataSource {
 
 	public InMemoryJSONDataSource(InputStream inputStream) {
 		super(inputStream);
-		scanner = createScanner(inputStream);
-		parse();
-	}
-	
-	public InMemoryJSONDataSource(String filePath) {
-		super(filePath);
-		parse();
 	}
 
-	private void parse() {
+	public InMemoryJSONDataSource(String filePath) {
+		super(filePath);
+	}
+
+	@Override
+	protected void parse() {
 		StringBuilder jsonContent = new StringBuilder();
 		while (scanner.hasNextLine()) {
 			jsonContent.append(scanner.nextLine());
@@ -73,61 +63,5 @@ public class InMemoryJSONDataSource extends AbstractFileSource implements InMemo
 				fieldsMap.put(key, String.valueOf(value));
 			}
 		}
-	}
-
-	@Override
-	public void open() {
-		if (opened) {
-			throw new IllegalStateException("DataSource is already open");
-		}
-		mapIterator = fieldsMap.keySet().iterator();
-		opened = true;
-	}
-
-	@Override
-	public String[] nextRow() {
-		checkIfOpen();
-		if (mapIterator.hasNext()) {
-			String key = mapIterator.next();
-			String value = fieldsMap.get(key);
-			return new String[] { key, value };
-		}
-		return null;
-	}
-
-	@Override
-	public boolean hasMoreData() {
-		checkIfOpen();
-		return mapIterator.hasNext();
-	}
-
-	@Override
-	public void reset() {
-		checkIfOpen();
-		mapIterator = fieldsMap.keySet().iterator();
-	}
-
-	public Iterator<String[]> iterator() {
-		return new Iterator<>() {
-			@Override
-			public boolean hasNext() {
-				return hasMoreData();
-			}
-
-			@Override
-			public String[] next() {
-				return nextRow();
-			}
-		};
-	}
-
-	@Override
-	public String[] getColumnNames() {
-		return fieldsMap.keySet().toArray(new String[] {});
-	}
-
-	@Override
-	public List<String[]> readAll() {
-		return super.readAll();
 	}
 }
